requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.0
jsonschema==4.22.0
Werkzeug==3.0.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-cov==5.0.0
pylint==3.2.3
